requirements.txt
fastapi==0.97.0
matplotlib==3.7.1
numpy==1.24.2
openai==0.27.8
opencv_python==4.1.2.30
opencv_python_headless==4.7.0.72
Pillow==9.4.0
Pillow==7.0.0
Pillow==9.5.0
Requests==2.31.0
uvicorn==0.22.0
